pkg/trace/api: add support for Unix Domain Sockets
This change adds support for Unix Domain Sockets by means of a
configuration setting. By default, this feature is disabled. To enable
it, simply point the yaml setting to the right socket path, for example:

    apm_config:
        receiver_socket: /tmp/trace.sock

This will enable the HTTP server to listen on unix:///tmp/trace.sock

The DD_APM_RECEIVER_SOCKET environment variable is an alias for the
above.